New solver logic for the t3 engine

+++ src/evaluator/map_evaluator.rs
@@ -4,11 +4,11 @@ use super::Evaluator;
 
 pub struct MapEvaluator {
     evaluator: Box<dyn Evaluator>,
-    mapper: Box<dyn Fn(i32) -> i32>,
+    mapper: Box<dyn Fn(f32) -> f32>,
 }
 
 impl MapEvaluator {
-    pub fn new(evaluator: Box<dyn Evaluator>, mapper: Box<dyn Fn(i32) -> i32>) -> Self {
+    pub fn new(evaluator: Box<dyn Evaluator>, mapper: Box<dyn Fn(f32) -> f32>) -> Self {
         Self { evaluator, mapper }
     }
 }
@@ -20,7 +20,7 @@ impl Evaluator for MapEvaluator {
         x: usize,
         y: usize,
         player: &crate::parser::Player,
-    ) -> i32 {
+    ) -> f32 {
         (self.mapper)(self.evaluator.score(board_state, x, y, player))
     }
 }
